% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.NucleicAcids.R
\name{connect.nucleic}
\alias{connect.nucleic}
\alias{connect.nucleic.pdb}
\alias{connect.nucleic.atoms}
\title{Connects atoms in Nucleic acids}
\usage{
\method{connect}{nucleic}(x, ...)

\method{connect.nucleic}{pdb}(x, ...)

\method{connect.nucleic}{atoms}(x, ...)
}
\arguments{
\item{x}{a \code{pdb} of \code{atoms} object containing atomic identifiers.}

\item{\dots}{arguments passed to methods.}
}
\value{
\code{connect.nucleic} returns an object of class \code{connect} containing
  the explicit connections between all nucleotide atoms.
Note: the detailed format of the returned value may change in the future.
}
\description{
Builds the skeleton for a nucleic acid, by connecting the individual atoms
for each nucleotide.
}
\details{
\code{connect.nucleic} is a generic function with methods for
  \code{pdb} and \code{atoms} objects.
It creates object of class \sQuote{connect} with the skeleton of each nucleotide.
}
\examples{
# Download some pdb files:
#   7EQJ = tRNA Val;
# x = read.pdb("9J4O.pdb")
# visualize(atoms(x), connect = connect.nucleic(x), type="l", pbc = FALSE);
}
\seealso{
\code{\link{pdb}}, \code{\link{connect}}
}
