% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorHelpers.R
\name{mirrorHelpers}
\alias{mirrorHelpers}
\alias{Mxy}
\alias{Mxy.coords}
\alias{Mxy.pdb}
\alias{Myz}
\alias{Myz.coords}
\alias{Myz.pdb}
\alias{Mzx}
\alias{Mzx.coords}
\alias{Mzx.pdb}
\alias{Mab}
\alias{Mab.coords}
\alias{Mab.pdb}
\alias{Mbc}
\alias{Mbc.coords}
\alias{Mbc.pdb}
\alias{Mca}
\alias{Mca.coords}
\alias{Mca.pdb}
\title{Helper Functions for reflection of Atomic Coordinates}
\usage{
Mxy(...)

\method{Mxy}{coords}(x, mask = TRUE, crystal = NULL, ...)

\method{Mxy}{pdb}(x, mask = TRUE, crystal = x$crystal, ...)

Myz(...)

\method{Myz}{coords}(x, mask = TRUE, crystal = NULL, ...)

\method{Myz}{pdb}(x, mask = TRUE, crystal = x$crystal, ...)

Mzx(...)

\method{Mzx}{coords}(x, mask = TRUE, crystal = NULL, ...)

\method{Mzx}{pdb}(x, mask = TRUE, crystal = x$crystal, ...)

Mab(...)

\method{Mab}{coords}(x, crystal, mask = TRUE, ...)

\method{Mab}{pdb}(x, crystal = x$crystal, mask = TRUE, ...)

Mbc(...)

\method{Mbc}{coords}(x, crystal, mask = TRUE, ...)

\method{Mbc}{pdb}(x, crystal = x$crystal, mask = TRUE, ...)

Mca(...)

\method{Mca}{coords}(x, crystal, mask = TRUE, ...)

\method{Mca}{pdb}(x, crystal = x$crystal, mask = TRUE, ...)
}
\arguments{
\item{\dots}{further arguments passed to or from other methods.}

\item{x}{an R object containing atomic coordinates.}

\item{mask}{a logical vector indicating the set of coordinates to which to
apply the reflection.}

\item{crystal}{an object of class \sQuote{crystal} used to convert fractional
into Cartesian coordinates (when needed).}
}
\value{
An object of the same class as \code{x} with reflected coordinates.
}
\description{
Reflection of atomic coordinates with respect to a specific Cartesian plane or
a plane defined by two lattice vectors.
}
\details{
These functions are helper functions to perform a reflection with respect to 
a specific Cartesian plan or a plan defined by two lattice vectors. All of 
them call the \code{mirror} function.
}
\examples{
# First lets read a pdb file
x <- read.pdb(system.file("examples/PCBM_ODCB.pdb", package="Rpdb"))
visualize(x, mode = NULL)

# Mirror operation with respect to the ab-plane
visualize(Mab(x), mode = NULL)
# Mirror operation with respect to the ab-plane for residue 1
visualize(Mab(x, mask = x$atoms$resid == 1), mode = NULL)

}
\seealso{
\code{\link{mirror}} and \code{\link{xyz2abc}},
\code{\link{abc2xyz}} for passing from Cartesian to fractional coordinates
(or Vice Versa).
}
\keyword{manip}
