% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\alias{select.pdb}
\alias{select.atoms}
\title{Select atoms}
\usage{
select(x, id, ...)

\method{select}{pdb}(x, id, residue, chain, ...)

\method{select}{atoms}(x, id, residue, chain, ...)
}
\arguments{
\item{x}{an R object containing atoms;}

\item{id}{an integer vector specifying the id of residues.}

\item{\dots}{arguments passed to methods.}

\item{residue}{a character vector specifying the residues to select.}

\item{chain}{character vector specifying which chains to process;}
}
\value{
\code{select} returns a vector with the element IDs.
}
\description{
Select groups of atoms from a PDB object.
}
\details{
\code{select} is a generic function, which enables
  the selection of element IDs from molecules.
}
