% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.Tools.R
\name{translate}
\alias{translate}
\alias{trans}
\title{Graphics Tools}
\usage{
trans(x, y = 0, z = 0, auto = FALSE)

translate(x, y = 0, z = 0, auto = FALSE)
}
\arguments{
\item{x, y, z}{amount to shift the observer in the x, y and z directions.}

\item{auto}{If TRUE, the location will be set automatically by RGL
to make the whole bounding box visible.
Unfortunately, it may fail, as it seems there are some bugs in rgl.}
}
\value{
\code{translate} returns invisibly the old position of the observer.
}
\description{
Tools to modify or set \code{rgl} objects, scenes and the observer.
\code{translate} performs a translation of the observer to a new position
  specified using relative shifts.
\code{trans} is an abbreviated alias to function \code{translate}.
}
