% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centres.roll.R
\name{centres.pp.roll}
\alias{centres.pp.roll}
\alias{centres.pp.roll.default}
\alias{centres.pp.roll.atoms}
\title{Centres-of-Geometry: Rolling Window}
\usage{
centres.pp.roll(x, window, ...)

\method{centres}{pp.roll.default}(x, window = 34, na.rm = TRUE, ...)

\method{centres.pp.roll}{atoms}(x, window = 34, chain = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{x}{an R object containing atomic coordinates.}

\item{window}{size of the rolling window, specified as number of amino-acids;}

\item{\dots}{further arguments passed to or from other methods.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}

\item{chain}{apply only to the respective chains, by default all chains;}
}
\value{
Return an object of class \sQuote{coords} containing the coordinates
  of the centres.
}
\description{
Computes centres-of-geometry using a rolling window on a polypeptide chain
}
\details{
\sQuote{centres.pp.roll} is a generic function to compute the centres-of-geometry
using a rolling window on an object containing atomic coordinates for
a polypeptide chain.
}
\seealso{
\code{\link{coords}}, \code{\link{atoms}}, \code{\link{pdb}}.
}
