% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{chains}
\alias{chains}
\alias{chains.atoms}
\alias{chains.pdb}
\title{Extract Chain-Codes from a PDB Molecule
Extract the codes of all chains from a PDB object.
\code{chains} is a generic function to extract the ids
of the chains from a PDB molecule.}
\usage{
chains(x, ...)

\method{chains}{atoms}(x, sorted = TRUE, ...)

\method{chains}{pdb}(x, ...)
}
\arguments{
\item{x}{Object of class \code{pdb} or class \code{atoms};}

\item{\dots}{arguments passed to other methods.}

\item{sorted}{logical indicating if the returned chain-codes should be sorted;}
}
\value{
\code{chains} returns the codes of the individual chains in the molecule.
}
\description{
Extract Chain-Codes from a PDB Molecule
Extract the codes of all chains from a PDB object.
\code{chains} is a generic function to extract the ids
of the chains from a PDB molecule.
}
