% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\alias{connect.default}
\alias{conect}
\alias{conect.default}
\alias{conect.coords}
\alias{conect.pdb}
\alias{is.connect}
\title{Create \sQuote{connect} Object}
\usage{
connect.default(eleid.1, eleid.2, ...)

conect(...)

\method{conect}{default}(eleid.1, eleid.2, ...)

\method{conect}{coords}(x, radii = 0.75, safety = 1.2, by.block = FALSE, ...)

\method{conect}{pdb}(x, safety = 1.2, by.block = FALSE, ...)

is.connect(x)
}
\arguments{
\item{eleid.1}{a integer vector containing the IDs of bonded atoms.}

\item{eleid.2}{a integer vector containing the IDs of bonded atoms.}

\item{\dots}{arguments passed to methods.}

\item{x}{an R object containing atomic coordinates.}

\item{radii}{a numeric vector containing atomic radii used to find neigbours.}

\item{safety}{a numeric value used to extend the atomic radii.}

\item{by.block}{a logical value indicating whether the connectivity has to be
determine by block (see details).}
}
\value{
\code{connect} returns a two-column data.frame of class
\sQuote{connect} whose rows contain the IDs of bonded atoms. The columns of
this data.frame are described below: \item{eleid.1}{a integer vector
containing the elements IDs defining the connectivity of the system.} 
\item{eleid.2}{a integer vector containing the elements IDs defining the
connectivity of the system.} \cr\cr \code{is.connect} returns TRUE if \code{x}
is an object of class \sQuote{coords} and FALSE otherwise.
}
\description{
Creates an object of class \sQuote{connect} containing the IDs of bonded atoms
defining the connectivity of a molecular system.
}
\details{
\code{connect} is a generic function to create objects of class 
\sQuote{connect}. The purpose of this class is to store CONNECT records from 
PDB files, indicating the connectivity of a molecular system.\cr The default 
method creates a \code{connect} object from its different components, i.e.: 
\code{eleid.1} and \code{eleid.2}. Both arguments have to be specified.\cr 
The S3 method for object of class \sQuote{coords} determine the connectivity 
from atomic coordinates. A distance matrix is computed, then, for each pair 
of atom the distance is compared to a bounding distance computed from atomic 
radii. If this distance is lower than the bounding distance then the atoms 
are assumed to be connected.\cr The S3 method for object of class 
\sQuote{pdb} first use element names to search for atomic radii in the 
\code{elements} data set. Then atomic coordinates and radii are passed to 
\code{conect.coords}.\cr If \code{by.block == TRUE}, a grid is defined to 
determined the connectivity by block. The method is slow but allow to deal 
with very large systems. \cr \code{is.connect} tests if an object of class 
\sQuote{connect}, i.e. if it has a \dQuote{class} attribute equal to 
\code{connect}.
}
\examples{
# If atom 1 is connected to atom 2, 3, 4 and 5
# then we can prepare the following 'connect' object:
x <- conect(rep(1,4), 2:5)
print(x)
is.connect(x)

# Compute connectivity from coordinates
x <- read.pdb(system.file("examples/PCBM_ODCB.pdb", package="Rpdb"), CONNECT = FALSE)
x$connect
x$connect <- conect(x)
x$connect

}
\seealso{
\code{\link{pdb}}
}
\keyword{classes}
