% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorialOperations.R
\name{vectorialOperations}
\alias{vectorialOperations}
\alias{dotProd}
\alias{vectNorm}
\alias{rotVect}
\alias{vectProd}
\title{Basic Vectorial Operations}
\usage{
dotProd(U, V)

vectNorm(U)

rotVect(U, n = 1)

vectProd(U, V)
}
\arguments{
\item{U}{a numeric vector of length 3.}

\item{V}{a numeric vector of length 3.}

\item{n}{an integer.}
}
\value{
\itemize{
\item{\code{dotProct} return a single element numeric vector.}
\item{\code{vectNorm} return a single element numeric vector.}
\item{\code{rotVect} return a numeric vector of length 3.}
\item{\code{vectProct} return a numeric vector of length 3.}
}
}
\description{
Basic vectorial operations such as scalar product and vectorial product
}
\examples{
Vx <- c(3,0,0)
vectNorm(Vx)
Vx <- Vx / vectNorm(Vx)
Vy <- c(0,1,0)
Vz <- vectProd(Vx, Vy)
print(Vz)

}
\seealso{
\code{\link[base]{matmult}}
}
\keyword{manip}
